Add optional npm-tag parameter to action (#30)
* Add optional npm-tag parameter to action to customize npm publish tag

The new npm-tag parameter allows to customize the npm publish tag. If omitted, the action defaults it to "latest". The YARN_NPM_TAG environment variable is passed to the bash script where it's used to customize the npm publish tag parameter.

* Add a step to set up Node.js with the version specified in the ".nvmrc" file

This commit adds a step to set up Node.js in the Github Action workflow with the version specified in the ".nvmrc" file. This ensures that the tests are run using the specified version of Node.js.