Run linting and testing GitHub workflows on PRs (#62)
Currently, code is not automatically linted and tests are not
automatically run on pull requests. This happens because the workflows
which would perform these steps are not configured to run on pull
requests but only on a push to `main`. This PR modifies the workflows to
run on the `pull_request` event.