Don't fail overall if Slack announcement fails
Typically, for a project, this action is run twice: once in dry-run mode
(to allow approvers to review the package before it is published) and
another time to issue the approval step and then actually publish the
package.

When the action is run in dry-run mode, an announcement on Slack is
created in order to notify approvers. If this fails in any way, however,
it halts the release workflow completely and prevents the second
instance of the action from running. This means that the package can
never be released automatically and must be released manually.

This commit fix this problem by instructing GitHub Actions that it is
not mandatory for the Slack announcement step to pass in order for the
release workflow to continue.